# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tome {
  id: number;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  tags?: string[];
  /** @nullable */
  sourceUrl?: string | null;
  featured: boolean;
  viewCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface TomeInput {
  /** @minLength 1 */
  title: string;
  excerpt?: string;
  /** @minLength 1 */
  content: string;
  category: string;
  tags?: string[];
  sourceUrl?: string;
  featured?: boolean;
}

export interface TomeUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  tags?: string[];
  sourceUrl?: string;
  featured?: boolean;
}

export interface TomeList {
  tomes: Tome[];
  total: number;
}

export interface Category {
  name: string;
  count: number;
}

export interface TomeStats {
  totalTomes: number;
  totalCategories: number;
  totalViews: number;
  recentCount: number;
  topCategories?: Category[];
}

export type OracleQueryContextItem = {
  title?: string;
  url?: string;
  content?: string;
};

export type OracleQueryHistoryItem = {
  role?: string;
  content?: string;
};

export interface OracleQuery {
  question: string;
  context?: OracleQueryContextItem[];
  history?: OracleQueryHistoryItem[];
}

export interface OracleResponse {
  answer: string;
}

export interface OracleLog {
  id: number;
  question: string;
  answer: string;
  createdAt: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  token: string;
  expiresAt: string;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminDashboard {
  totalTomes: number;
  totalUsers: number;
  totalOracleQueries: number;
  totalViews?: number;
  recentTomes?: Tome[];
  recentActivity: ActivityItem[];
  categoryBreakdown?: Category[];
}

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
  tomeCount?: number;
}

export type ListTomesParams = {
category?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListRecentTomesParams = {
limit?: number;
};

export type ListOracleLogsParams = {
limit?: number;
offset?: number;
};

export type ListUsersParams = {
limit?: number;
offset?: number;
};