# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}