# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  token: string;
  expiresAt: string;
}