# lib/api-zod/src/generated/types/oracleLog.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface OracleLog {
  id: number;
  question: string;
  answer: string;
  createdAt: string;
}