# lib/api-zod/src/generated/types/oracleQuery.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */
import type { OracleQueryContextItem } from './oracleQueryContextItem';
import type { OracleQueryHistoryItem } from './oracleQueryHistoryItem';

export interface OracleQuery {
  question: string;
  context?: OracleQueryContextItem[];
  history?: OracleQueryHistoryItem[];
}