# lib/api-zod/src/generated/types/tome.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface Tome {
  id: number;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  tags?: string[];
  /** @nullable */
  sourceUrl?: string | null;
  featured: boolean;
  viewCount?: number;
  createdAt: string;
  updatedAt?: string;
}