# lib/api-zod/src/generated/types/tomeInput.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface TomeInput {
  /** @minLength 1 */
  title: string;
  excerpt?: string;
  /** @minLength 1 */
  content: string;
  category: string;
  tags?: string[];
  sourceUrl?: string;
  featured?: boolean;
}