# lib/api-zod/src/generated/types/tomeList.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */
import type { Tome } from './tome';

export interface TomeList {
  tomes: Tome[];
  total: number;
}