# lib/api-zod/src/generated/types/tomeStats.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';

export interface TomeStats {
  totalTomes: number;
  totalCategories: number;
  totalViews: number;
  recentCount: number;
  topCategories?: Category[];
}