# lib/api-zod/src/generated/types/tomeUpdate.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface TomeUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  category?: string;
  tags?: string[];
  sourceUrl?: string;
  featured?: boolean;
}