# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Akashic Record API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  createdAt: string;
  tomeCount?: number;
}